+++ lib/impulse_player.dart
@@ -1,5 +1,3 @@
-import 'dart:ffi';
-
 import 'package:flutter/material.dart';
 import 'package:impulse_player_flutter/plugin/plugin_platform.dart';